El parlante metia 80 ms de silencio adentro de cada primera palabra
El usuario reporto que a veces se le corta la voz. De eso no habia en toda la
bitacora una sola marca: la salida reproduce con ReadFully, que es lo correcto
--devolver cero muestras hace que WinMM de la reproduccion por terminada-- pero
cuando la cola se queda corta el hueco se rellena con silencio y no lo denuncia
nadie. Ni excepcion, ni contador, ni renglon.

Lo primero fue medirlo, con la geometria real del parlante y NAudio de verdad.
Cuatro corridas identicas:

  en la cola al dar Play:   20 ms -> 80 ms de relleno
                            40 ms -> 60 ms
                            60 ms -> 40 ms
                            80 ms -> 20 ms
                           100 ms ->  0 ms
                           120 ms ->  0 ms

NAudio le pide al proveedor TODOS sus buferes en la primera vuelta de su hilo:
cien milisegundos de una. La reproduccion arrancaba con el primer bloque de
veinte que llegara, asi que ochenta de esos cien salian de silencio, adentro de
la primera palabra, en cada respuesta. El relleno es exactamente la latencia del
driver menos lo encolado, clavado, y cero justo al llegar a la latencia.

De ahi sale el colchon, y de ahi sale que sea de 100 y no de 250: todo lo que se
espere por encima de lo medido es demora pura, y se paga tambien en la respuesta
que sigue a una interrupcion, que es la interaccion que el usuario dijo que
andaba bien. El techo de lo que todo esto puede demorar una voz son 150 ms.

Probado contra el peor riesgo del cambio --que la deje muda-- con el parlante de
verdad y tarjeta de verdad, incluidos los dos casos peligrosos:

  una palabra sola (20 ms, el turno cierra)      403 ms   sono
  respuesta normal (100 ms de una)               137 ms   sono
  a cuentagotas (20 ms y nada mas, sin cerrar)   288 ms   sono
  goteo lento (20 ms cada 60 ms, sin cerrar)     380 ms   sono

Lo demas es instrumento, y va con tres defectos corregidos que lo hacian mentir:

  - El reparto se contaba desde el Play y no desde la primera lectura. Entre uno
    y otra hay un despacho del ThreadPool que se midio tardando 56 ms: se
    informaba un atraso que no sono, porque no se le habia entregado nada al
    dispositivo todavia. Un falso positivo en cada arranque.

  - Una herramienta lenta a mitad de turno seca la cola con el turno abierto, y
    se informaba como un hueco del tamano de la herramienta: los 10,1 s que hay
    en la bitacora del usuario habrian entrado como un corte de voz, enterrando
    a los de verdad. Ahora un silencio de mas de un segundo no es un tajo en la
    voz sino que dejo de hablar, y el final de la respuesta borra el hueco
    abierto en vez de que el hilo del driver consulte una bandera.

  - El aviso de atraso se enclavaba despues del primero y la charla entera
    quedaba muda. El reparto es acumulado y su techo es la latencia del driver,
    asi que una vez ido a negativo no vuelve. Va por marca de agua: sostenido se
    informa una sola vez, uno nuevo encima del anterior se informa por lo que
    agrego.

Y dos avisos que faltaban, que son candidatos directos al sintoma: un corte de
transporte vacia la cola del parlante --desde el cuarto se oye igual que la voz
cortandose a mitad de frase-- y hasta ahora se tragaba entero; y la rafaga de la
antesala del microfono, para poder cruzarla con los cortes.

846 verdes, 0 omitidas.

Sin diagnosticar todavia: por que se corta. Esto no lo arregla, lo hace visible.
El renglon "vivo.parlante" al cerrar la charla es lo primero que hay que mirar.

+++ src/Viernes.Core/Live/LiveAudioFormat.cs
@@ -52,7 +52,11 @@ public static class LiveAudioFormat
             : milliseconds * InputSampleRate / 1000 * BytesPerSample;
 
     /// <summary>Cuánto dura, dicho, este bloque de audio de respuesta.</summary>
-    public static TimeSpan OutputDurationOf(int byteCount) =>
+    /// <remarks>
+    /// Cuenta en <c>long</c> y no en <c>int</c> porque no siempre es un bloque: <c>LivePlayout</c> le
+    /// pasa todo lo que se le entregó al driver desde que arrancó una respuesta, y eso se acumula.
+    /// </remarks>
+    public static TimeSpan OutputDurationOf(long byteCount) =>
         byteCount <= 0
             ? TimeSpan.Zero
             : TimeSpan.FromSeconds((double)byteCount / (OutputSampleRate * BytesPerSample));
